refactor: move dateFormatStringOption and datePattern to option_date.go, extract option_date_test.go
