+++ internal/core/option_date.go
@@ -1,5 +1,12 @@
 package core
 
+import (
+	"fmt"
+	"regexp"
+)
+
+var datePattern = regexp.MustCompile(`^\d{4}-\d{2}-\d{2}$`)
+
 func (s *OptionService) WithCreatedSince(date string) RequestOption {
 	return dateFormatStringOption(ParamCreatedSince, date)
 }
@@ -42,4 +49,19 @@ func (s *OptionService) WithDueDateUntil(date string) RequestOption {
 
 func (s *OptionService) WithReleaseDueDate(date string) RequestOption {
 	return dateFormatStringOption(ParamReleaseDueDate, date)
+}
+
+// dateFormatStringOption builds a RequestOption that validates the string matches
+// "yyyy-MM-dd" format before setting it.
+func dateFormatStringOption(paramType APIParamOptionType, date string) RequestOption {
+	return &APIParamOption{
+		Type: paramType,
+		CheckFunc: func() error {
+			if !datePattern.MatchString(date) {
+				return NewValidationError(fmt.Sprintf("%s must be formatted as yyyy-MM-dd, got %q", paramType.Value(), date))
+			}
+			return nil
+		},
+		SetFunc: setStringFunc(paramType, date),
+	}
 }
